Unset the query data after it ran successfully
As issue #1 suggests, remove the where binds after the query ran successfuly, because they are concatenated on subsequent queries. To do this properly a new protected method **_runQuery** was added, that *shuts down* the query, meaning in unsets all, join, order by, group by, and limit clauses as well as the where and whereBinds arrays. This was ninja edited on github, so it might break something, take extra care.